optimized fnv hash functions and fixed tests

+++ hashmap.h
@@ -14,7 +14,7 @@ extern "C" {
 
 #define DEFAULT_MAP_SIZE 32
 
-typedef unsigned long hash_t;
+typedef size_t hash_t;
 
 typedef struct hashmap
 {
